requirements.txt
openai
langchain
setuptools-rust
openai-whisper
pysubs2
Flask
python-dotenv
requests
